refactor(playwright): fix e2e-ct action paths, use default delay when unbolding text
