+++ .github/workflows/e2e-ct.yml
@@ -78,7 +78,7 @@ jobs:
         if: ${{ always() && steps.parse-playwright-report.conclusion == 'success' && github.event_name == 'pull_request' }}
         with:
           comment_tag: 'playwright-ct-report'
-          filePath: ${{ github.workspace }}/packages/sanity/playwright-report-output.md
+          filePath: ${{ github.workspace }}/playwright-report-output.md
 
       - uses: actions/upload-artifact@v3
         if: always()
